refactor: move finalizeSmitheryConnection logic to API route
The finalize logic mixed two concerns: checking Smithery connection status
and updating Prisma. Move the database update directly into the API route
where it belongs, keeping the wrapper file focused on provider delegation.

This removes ~35 lines from the wrapper and makes the data flow clearer:
API route → provider.getConnectionStatus() → prisma.update()
